+++ packages/react/src/components/EmptyApp/src/components/EmptyAppContentImageElement/index.ts
@@ -1,0 +1,4 @@
+import { EmptyAppContentImageElement } from "./EmptyAppContentImageElement";
+
+export { EmptyAppContentImageElement } from "./EmptyAppContentImageElement";
+export default EmptyAppContentImageElement;