+++ packages/react/src/components/EmptyApp/src/components/EmptyAppContentSection/index.ts
@@ -1,0 +1,4 @@
+import { EmptyAppContentSection } from "./EmptyAppContentSection";
+
+export { EmptyAppContentSection } from "./EmptyAppContentSection";
+export default EmptyAppContentSection;