+++ packages/react/src/components/EmptyApp/src/components/EmptyAppContentTextElement/emptyAppContentTextElement.types.ts
@@ -1,0 +1,11 @@
+import { HTMLAttributes } from "react";
+
+export interface EmptyAppContentTextElementProperties {
+  title: string;
+  content: string;
+  bottomDivider?: boolean;
+}
+
+export type EmptyAppContentTextElementProps =
+  EmptyAppContentTextElementProperties &
+    Omit<HTMLAttributes<HTMLElement>, "color">;