+++ packages/react/src/components/EmptyApp/src/components/EmptyAppHeroSection/emptyAppHeroSection.types.ts
@@ -1,0 +1,14 @@
+import { ReactNode } from "react";
+import { BoxProps, ThumbnailProps } from "@nimbus-ds/components";
+
+export interface EmptyAppHeroSectionProperties
+  extends Pick<ThumbnailProps, "src" | "alt"> {
+  title: string;
+  titleChildren?: ReactNode;
+  subtitle?: string;
+  content: string;
+  actions?: ReactNode;
+}
+
+export type EmptyAppHeroSectionProps = EmptyAppHeroSectionProperties &
+  Omit<BoxProps, "backgroundImage" | "minHeight" | "pt">;