+++ packages/react/src/components/EmptyApp/src/components/EmptyAppHeroSection/index.ts
@@ -1,0 +1,4 @@
+import { EmptyAppHeroSection } from "./EmptyAppHeroSection";
+
+export { EmptyAppHeroSection } from "./EmptyAppHeroSection";
+export default EmptyAppHeroSection;