+++ packages/react/src/components/EmptyApp/src/components/EmptyAppPlanCard/emptyAppPlanCard.types.ts
@@ -1,0 +1,10 @@
+import { HTMLAttributes, ReactNode } from "react";
+
+export interface EmptyAppPlanCardProperties {
+  planContext: string;
+  planName: string;
+  children: ReactNode;
+}
+
+export type EmptyAppPlanCardProps = EmptyAppPlanCardProperties &
+  HTMLAttributes<HTMLElement>;