+++ packages/react/src/components/EmptyApp/src/components/EmptyAppPlanItem/emptyAppPlanItem.types.ts
@@ -1,0 +1,9 @@
+import { BoxProps } from "@nimbus-ds/components";
+
+export interface EmptyAppPlanItemProperties {
+  label: string;
+  disabled?: boolean;
+}
+
+export type EmptyAppPlanItemProps = EmptyAppPlanItemProperties &
+  Omit<BoxProps, "display" | "alignItems" | "gap">;