+++ packages/react/src/components/EmptyApp/src/index.ts
@@ -1,0 +1,5 @@
+import { EmptyApp } from "./EmptyApp";
+
+export { EmptyApp } from "./EmptyApp";
+export type { EmptyAppProps } from "./emptyApp.types";
+export default EmptyApp;