+++ packages/react/CHANGELOG.md
@@ -2,6 +2,30 @@
 
 Nimbus is an open-source Design System created by Tiendanube / Nuvesmhop’s team to empower and enhance more stories every day, with simplicity, accessibility, consistency and performance.
 
+## 2024-03-06 `1.8.0`
+
+### 🎉 New features
+
+- Added new `EmptyApp` component. ([#95](https://github.com/TiendaNube/nimbus-patterns/pull/95) by [@juanchigallego](https://github.com/juanchigallego))
+- Added new `EmptyAppContentImageElement` subcomponent. ([#95](https://github.com/TiendaNube/nimbus-patterns/pull/95) by [@juanchigallego](https://github.com/juanchigallego))
+- Added new `EmptyAppContentSection` subcomponent. ([#95](https://github.com/TiendaNube/nimbus-patterns/pull/95) by [@juanchigallego](https://github.com/juanchigallego))
+- Added new `EmptyAppContentTextElement` subcomponent. ([#95](https://github.com/TiendaNube/nimbus-patterns/pull/95) by [@juanchigallego](https://github.com/juanchigallego))
+- Added new `EmptyAppHeroSection` subcomponent. ([#95](https://github.com/TiendaNube/nimbus-patterns/pull/95) by [@juanchigallego](https://github.com/juanchigallego))
+- Added new `EmptyAppPlanCard` subcomponent. ([#95](https://github.com/TiendaNube/nimbus-patterns/pull/95) by [@juanchigallego](https://github.com/juanchigallego))
+- Added new `EmptyAppPlanItem` subcomponent. ([#95](https://github.com/TiendaNube/nimbus-patterns/pull/95) by [@juanchigallego](https://github.com/juanchigallego))
+- Added `children` property to `EmptyApp` component API. ([#95](https://github.com/TiendaNube/nimbus-patterns/pull/95) by [@juanchigallego](https://github.com/juanchigallego))
+- Added `src` and `alt` properties to `EmptyAppContentImageElement` subcomponent API. ([#95](https://github.com/TiendaNube/nimbus-patterns/pull/95) by [@juanchigallego](https://github.com/juanchigallego))
+- Added `title`, `subtitle`, `actions` and `children` properties to `EmptyAppContentSection` subcomponent API. ([#95](https://github.com/TiendaNube/nimbus-patterns/pull/95) by [@juanchigallego](https://github.com/juanchigallego))
+- Added `title`, `content` and `bottomDivider` properties to `EmptyAppContentTextElement` subcomponent API. ([#95](https://github.com/TiendaNube/nimbus-patterns/pull/95) by [@juanchigallego](https://github.com/juanchigallego))
+- Added `src`, `alt`, `title`, `subtitle`, `content` and `actions` properties to `EmptyAppHeroSection` subcomponent API. ([#95](https://github.com/TiendaNube/nimbus-patterns/pull/95) by [@juanchigallego](https://github.com/juanchigallego))
+- Added `planContext`, `planName` and `children` properties to `EmptyAppPlanCard` subcomponent API. ([#95](https://github.com/TiendaNube/nimbus-patterns/pull/95) by [@juanchigallego](https://github.com/juanchigallego))
+- Added `label` and `disabled` properties to `EmptyAppPlanItem` subcomponent API. ([#95](https://github.com/TiendaNube/nimbus-patterns/pull/95) by [@juanchigallego](https://github.com/juanchigallego))
+
+### 📚 3rd party library updates
+
+- Updated `@nimbus-ds/styles@9.11.0`. ([#95](https://github.com/TiendaNube/nimbus-patterns/pull/95) by [@juanchigallego](https://github.com/juanchigallego))
+- Updated `@nimbus-ds/components@5.5.1`. ([#95](https://github.com/TiendaNube/nimbus-patterns/pull/95) by [@juanchigallego](https://github.com/juanchigallego))
+
 ## 2024-02-07 `1.7.8`
 
 ### 🎉 New features